Update calories left

+++ foodbody/FoodBody/Models/New Group/DailyLogModel.swift
@@ -11,7 +11,7 @@ import ObjectMapper
 
 class DailyLogModel: Mappable  {
     
-    var step: Double?
+    var step: Int?
     var calo_threshold: Double?
     var remain_calo: Double?
     var date: String?
